Extract MCP example-tools cluster out of mcp-server.ts
Starts Wave C (mcp-server.ts decomposition, the actual largest
apps/api source file at 5957 lines, untouched by the prior wave).
Moves the pure tool-result/BEHAVIOURAL_CONTRACT/SESSION_KEY_PROP
scaffolding into mcp-tool-support.ts, and the list_examples/
get_example/list_example_files/read_example_file cluster into
mcp-example-tools.ts as createExampleTools({resolveAuth,
injectChannel}).
