+++ eslint-rules/module-boundary-map.mjs
@@ -175,6 +175,8 @@ const FILE_BUCKET = {
   // agent-surface: channel + MCP + kit
   'agent-channel': 'agent-surface',
   'mcp-server': 'agent-surface',
+  'mcp-tool-support': 'agent-surface',
+  'mcp-example-tools': 'agent-surface',
   'mcp-server-discovery': 'agent-surface',
   'mcp-ui': 'agent-surface',
   'mcp-oauth-metadata': 'agent-surface',
